pyproject.toml
[tool.poetry]
name = "ezred2bot"
version = "0.1.0"
description = ""
authors = ["Valentin Alix"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
pydantic = "^2.7.4"
unidecode = "^1.3.8"
tqdm = "^4.66.4"
sortedcontainers = "^2.4.0"
mypy = "^1.10.0"
python-dotenv = "^1.0.1"
matplotlib = "^3.9.0"
cachetools = "^5.3.3"
numpy = "^1.8.0"
fastapi = "^0.111.0"
alembic = "^1.13.1"
psycopg2-binary = "^2.9.9"
aiohttp = "^3.9.5"
uvicorn = "^0.30.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
